exclude optical

+++ src/sql/install/base.ddl.sql
@@ -1106,7 +1106,7 @@ CREATE TABLE IF NOT EXISTS kandidaten (
     losnummer_stimmzettelgruppe INT DEFAULT 0,
     losnummer INT DEFAULT 0,
     ist_gewaehlt TINYINT DEFAULT 0,
-    bp_column integer not null default 1,
+    bp_column bigint not null default 1,
 
     PRIMARY KEY (id),
     CONSTRAINT fk_kandidaten_stimmzettelgruppen FOREIGN KEY (stimmzettelgruppen) REFERENCES `stimmzettelgruppen` (`ridx`) ON
